Integrate oegenerate to oeutil
