+++ tests/host_verify/host/host.cpp
@@ -335,9 +335,9 @@ int main()
     // Report only tests
     //
 
-    // These files are generated by oegenerate and do not always exists.
+    // These files are generated by oeutil and do not always exist.
     // Run these tests if the file exists.  The Jenkins CI/CD system
-    // is responsible for running oegenerate to generate these files.
+    // is responsible for running oeutil to generate these files.
     //
     if (_validate_file(CERT_EC_FILENAME, false))
         _verify_cert(CERT_EC_FILENAME, true);
