+++ tests/tools/oecert/README.md
@@ -1,0 +1,4 @@
+oecert
+=====
+
+This function has been moved from the current directory, `/tests/tools/oecert` and integrated to `/tools/oeutil`. Please refer to the [oeutil README file](/tools/oeutil/README.md) for more information.
