--- tests/tools/oegenerate/README.md
@@ -1,4 +1,0 @@
-oegenerate
-=====
-
-This tool has been moved from the current directory, `/tests/tools/oegenerate`, to `/tools/oegenerate`. Please refer to [the README file](/tools/oegenerate/README.md) for more information.
