+++ tools/CMakeLists.txt
@@ -65,7 +65,7 @@ if (OE_SGX)
   add_subdirectory(oesgx)
   add_subdirectory(oesign)
   add_subdirectory(oeverify)
-  if (BUILD_OEGENERATE_TOOL)
-    add_subdirectory(oegenerate)
+  if (BUILD_OEUTIL_TOOL)
+    add_subdirectory(oeutil)
   endif ()
 endif ()
