+++ tools/oeutil/README.md
@@ -1,7 +1,17 @@
-oegenerate
+OEUTIL
 =====
 
-`oegenerate` is a debugging utility that generates the following files in binary format, or dump them in readable text, and verify them:
+`oeutil` supports the following OE commands, which can be abbreviated to a prefix:
+
+    1. generate-evidence
+
+Usage: `oeutil <command> <options>`
+
+
+oeutil generate-evidence
+=====
+
+`oeutil generate-evidence` can generate the following files in binary format, or dump them in readable text, and verify them:
 
  1. Self-signed certificates (in der format) used for remote attestation over TLS.
  2. An OE report.
@@ -9,11 +19,11 @@ oegenerate
  4. An endorsement for OE report/evidence.
 
 For certificates, the user can pass in the public/private key.
-`oegenerate` is not suitable for production use.
+`oeutil generate-evidence` is not suitable for production use.
 
-Usage: `oegenerate Options`
+Usage: `oeutil generate-evidence <options>`
 
-where `Options` are:
+where `options` are:
 
     -f, --format <format_option>: generate evidence, a report, or a certificate, where format_option can be one of the following (case insensitive):
         cert <private_key> <public_key>: a remote attestation certificate in DER format.
@@ -25,25 +35,25 @@ where `Options` are:
     -o, --out <filename>: generate an output file for a remote attestation certificate, a report, or evidence.
     -e, --endorsements <filename>: output a report or evidence, and also its endorsements binary.
     -v, --verify: verify the generated remote attestation certificate, report, or evidence.
-    -l, --log <filename>: generate a log file (default: oegenerate.log).
+    -l, --log <filename>: generate a log file (default: oeutil_generate_evidence.log).
     --verbose: enable verbose output.
 
 Note that parameters are not case-sensitive.
 
 Example 1 Generate, verify and dump a ceritificate. Without "--out", there will be no certificate output file:
 
-    ./oegenerate --format cert keyecec.pem publickeyec.pem --verify --verbose
+    ./oeutil generate-evidence --format cert keyecec.pem publickeyec.pem --verify --verbose
 
 Example 2 Generate an OE report and output its endorsements binary to "endorsements.bin", and output OE report to "report.bin":
 
-    ./oegenerate --format legacy_report_remote --endorsements endorsements.bin --out report.bin
+    ./oeutil generate --format legacy_report_remote --endorsements endorsements.bin --out report.bin
 
 Example 3 Generate and verify OE evidence in SGX_ECDSA format, dump evidence buffer and its verified claims, and output OE evidence to "evidence.bin":
 
-    ./oegenerate --format sgx_ecdsa --verify --verbose --out evidence.bin
+    ./oeutil gen --format sgx_ecdsa --verify --verbose --out evidence.bin
 
 
-Creating RSA and EC keys pairs in linux using openssl.
+Creating RSA and EC keys pairs in Linux using openssl.
 Generate an RSA private key, of size 2048, and output it to a file named key.pem
 =====
 openssl genrsa -out keyrsa.pem 2048
