+++ tools/oeutil/enc/enc.conf
