+++ tools/oeutil/host/generate_evidence.cpp
@@ -1,16 +1,17 @@
 // Copyright (c) Open Enclave SDK contributors.
 // Licensed under the MIT License.
 
-#include "evidence.h"
-
 #include <ctype.h>
 #include <openenclave/attestation/sgx/evidence.h>
 #include <openenclave/attestation/sgx/report.h>
 #include <openenclave/attestation/verifier.h>
+#include <openenclave/corelibc/limits.h>
 #include <openenclave/host.h>
 #include <openenclave/internal/hexdump.h>
+#include <openenclave/internal/pem.h>
 #include <openenclave/internal/raise.h>
 #include <openenclave/internal/report.h>
+#include <openenclave/internal/sgx/tests.h>
 #include <openenclave/internal/sgxcertextensions.h>
 #include <openenclave/internal/tests.h>
 #include <openssl/bio.h>
@@ -19,7 +20,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "oegenerate_u.h"
+#if defined(__linux__)
+#include <unistd.h>
+#elif defined(_WIN32)
+#include <windows.h>
+#endif
+#include "oeutil_u.h"
 
 #if defined(__linux__)
 #include <dlfcn.h>
@@ -29,17 +35,71 @@
 
 #include "../../../common/attest_plugin.h"
 #include "../../../common/sgx/collateral.h"
+#include "../../../common/sgx/endorsements.h"
 #include "../../../common/sgx/quote.h"
 #include "../../../host/sgx/sgxquoteprovider.h"
-
+#include "generate_evidence.h"
+#include "parse_args_helper.h"
 extern FILE* log_file;
 
-#define OE_PEM_BEGIN_CERTIFICATE "-----BEGIN CERTIFICATE-----"
-#define OE_PEM_BEGIN_CERTIFICATE_LEN (sizeof(OE_PEM_BEGIN_CERTIFICATE) - 1)
-#define OE_PEM_END_CERTIFICATE "-----END CERTIFICATE-----"
-#define OE_PEM_END_CERTIFICATE_LEN (sizeof(OE_PEM_END_CERTIFICATE) - 1)
+#define DEFAULT_LOG_FILE "oeutil_generate_evidence.log"
+#define ENCLAVE_FILENAME_SUFFIX "_enc.signed"
+#define INPUT_PARAM_OPTION_FORMAT "--format"
+#define INPUT_PARAM_OPTION_CERT "cert"
+#define INPUT_PARAM_OPTION_ENDORSEMENTS_FILENAME "--endorsements"
+#define INPUT_PARAM_OPTION_QUOTE_PROC "--quote-proc"
+#define INPUT_PARAM_QUOTE_IN_PROC "in"
+#define INPUT_PARAM_QUOTE_OUT_OF_PROC "out"
+#define INPUT_PARAM_OPTION_VERIFY "--verify"
+#define INPUT_PARAM_OPTION_OUT_FILE "--out"
+#define INPUT_PARAM_OPTION_LOG_FILE "--log"
+#define INPUT_PARAM_OPTION_VERBOSE "--verbose"
+#define INPUT_PARAM_OPTION_HELP "--help"
+#define INPUT_PARAM_OPTION_LEGACY_REPORT_REMOTE "LEGACY_REPORT_REMOTE"
+#define INPUT_PARAM_OPTION_SGX_ECDSA "SGX_ECDSA"
+#define INPUT_PARAM_OPTION_SGX_EPID_LINKABLE "SGX_EPID_LINKABLE"
+#define INPUT_PARAM_OPTION_SGX_EPID_UNLINKABLE "SGX_EPID_UNLINKABLE"
+#define SHORT_INPUT_PARAM_OPTION_FORMAT "-f"
+#define SHORT_INPUT_PARAM_OPTION_ENDORSEMENTS_FILENAME "-e"
+#define SHORT_INPUT_PARAM_OPTION_QUOTE_PROC "-p"
+#define SHORT_INPUT_PARAM_OPTION_VERIFY "-v"
+#define SHORT_INPUT_PARAM_OPTION_OUT_FILE "-o"
+#define SHORT_INPUT_PARAM_OPTION_LOG_FILE "-l"
+#define SHORT_INPUT_PARAM_OPTION_HELP "-h"
+#define SGX_AESM_ADDR "SGX_AESM_ADDR"
+#if defined(_WIN32)
+#define SGX_AESM_ADDR_MAXSIZE 32
+#endif
+
 #define SGX_EXTENSION_OID_STR "1.2.840.113741.1.13.1"
 #define MAX_BUFFER_SIZE 65536
+
+// Static constants for evidence UUIDs
+static const oe_uuid_t _sgx_ecdsa_uuid = {OE_FORMAT_UUID_SGX_ECDSA};
+static const oe_uuid_t _sgx_epid_linkable_uuid = {
+    OE_FORMAT_UUID_SGX_EPID_LINKABLE};
+static const oe_uuid_t _sgx_epid_unlinkable_uuid = {
+    OE_FORMAT_UUID_SGX_EPID_UNLINKABLE};
+
+// Structure to store input parameters
+typedef struct _input_parameters
+{
+    const char* private_key_filename;
+    const char* public_key_filename;
+    const char* out_filename;
+    const char* log_filename;
+    const char* endorsements_filename;
+    const char* quote_proc;
+    bool generate_certificate;
+    bool generate_legacy_report_remote;
+    bool generate_sgx_ecdsa;
+    bool generate_sgx_epid_linkable;
+    bool generate_sgx_epid_unlinkable;
+    bool verify;
+    bool verbose;
+} input_parameters_t;
+
+static input_parameters_t _parameters;
 
 void log(const char* fmt, ...)
 {
@@ -60,6 +120,197 @@ void log(const char* fmt, ...)
     {
         printf("%s", message);
     }
+}
+
+OE_INLINE uint16_t read_uint16(const uint8_t* p)
+{
+    return (uint16_t)(p[0] | (p[1] << 8));
+}
+
+OE_INLINE uint32_t read_uint32(const uint8_t* p)
+{
+    return (uint32_t)(p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24));
+}
+
+static oe_result_t _read_key(const char* filename, uint8_t** data, size_t* size)
+{
+    FILE* fp;
+    fopen_s(&fp, filename, "rb");
+    size_t file_size;
+    oe_result_t result = OE_FAILURE;
+    uint8_t* memory = nullptr;
+
+    if (fp == nullptr)
+        goto done;
+
+    // Find file size
+    fseek(fp, 0, SEEK_END);
+    file_size = (size_t)ftell(fp);
+    fseek(fp, 0, SEEK_SET);
+
+    // Account for '\0'
+    memory = (uint8_t*)malloc(file_size + 1);
+    if (memory == nullptr)
+    {
+        printf("Failed to allocate memory.\n");
+        goto done;
+    }
+
+    if (fread(memory, 1, file_size, fp) == file_size)
+    {
+        memory[file_size] = '\0';
+        printf("Read in key: %s\n", memory);
+    }
+    else
+    {
+        printf("Failed to read in key: %s\n", filename);
+        goto done;
+    }
+
+    *data = memory;
+    *size = file_size + 1;
+    memory = nullptr;
+
+    result = OE_OK;
+
+done:
+    if (fp)
+        fclose(fp);
+    free(memory);
+
+    return result;
+}
+
+static void decode_crl_der(const uint8_t* data, size_t data_size)
+{
+    X509_CRL* x509;
+    BIO* input = BIO_new_mem_buf(data, (int)data_size);
+    x509 = d2i_X509_CRL_bio(input, NULL);
+    if (x509)
+        X509_CRL_print_fp(log_file, x509);
+    BIO_free_all(input);
+}
+
+static void decode_crl_pem(const uint8_t* data, size_t data_size)
+{
+    X509_CRL* x509;
+    BIO* input = BIO_new_mem_buf(data, (int)data_size);
+    x509 = PEM_read_bio_X509_CRL(input, NULL, NULL, NULL);
+    if (x509)
+        X509_CRL_print_fp(log_file, x509);
+    BIO_free_all(input);
+}
+
+// DCAP client (libdcap_quoteprov) log callback to this function.
+static void oeutil_quote_provider_log(
+    sgx_ql_log_level_t level,
+    const char* message)
+{
+    const char* level_string = level == 0 ? "ERROR" : "INFO";
+
+    log("[%s]: %s\n", level_string, message);
+}
+
+// Set DCAP client (libdcap_quoteprov) log callback
+static void set_log_callback()
+{
+#if defined(__linux__)
+    extern oe_sgx_quote_provider_t provider;
+
+    // Initialize quote provider and set log callback
+    oe_initialize_quote_provider();
+
+    sgx_ql_set_logging_function_t set_log_fcn =
+        (sgx_ql_set_logging_function_t)dlsym(
+            provider.handle, "sgx_ql_set_logging_function");
+    if (set_log_fcn != nullptr)
+    {
+        set_log_fcn(oeutil_quote_provider_log);
+    }
+#endif
+}
+
+oe_result_t output_file(
+    const char* file_name,
+    const uint8_t* data,
+    size_t data_size)
+{
+    FILE* output = nullptr;
+    fopen_s(&output, file_name, "wb");
+    if (!output)
+    {
+        log("Failed to open output file %s\n", file_name);
+        return OE_FAILURE;
+    }
+    fwrite(data, data_size, 1, output);
+    fclose(output);
+    return OE_OK;
+}
+
+static void _display_help(const char* command)
+{
+    printf(
+        "Generate-evidence Usage: %s generate-evidence "
+        "<options>\n",
+        command);
+    printf("options:\n");
+    printf(
+        "\t%s, %s <format_option>: generate evidence, a report, or a "
+        "certificate, where format_option can be one of the following (case "
+        "insensitive):\n",
+        SHORT_INPUT_PARAM_OPTION_FORMAT,
+        INPUT_PARAM_OPTION_FORMAT);
+    printf(
+        "\t\t%s <private_key> <public_key>: a remote attestation certificate "
+        "in DER format.\n",
+        INPUT_PARAM_OPTION_CERT);
+    printf(
+        "\t\t%s: a report in OE_FORMAT_UUID_LEGACY_REPORT_REMOTE format.\n",
+        INPUT_PARAM_OPTION_LEGACY_REPORT_REMOTE);
+    printf(
+        "\t\t%s: evidence in OE_FORMAT_UUID_SGX_ECDSA format.\n",
+        INPUT_PARAM_OPTION_SGX_ECDSA);
+    printf(
+        "\t\t%s: evidence in OE_FORMAT_UUID_SGX_EPID_LINKABLE format.\n",
+        INPUT_PARAM_OPTION_SGX_EPID_LINKABLE);
+    printf(
+        "\t\t%s: evidence in OE_FORMAT_UUID_SGX_EPID_UNLINKABLE format.\n",
+        INPUT_PARAM_OPTION_SGX_EPID_UNLINKABLE);
+    printf(
+        "\t%s, %s <in|out>: use SGX in-process or out-of-process quoting.\n",
+        SHORT_INPUT_PARAM_OPTION_QUOTE_PROC,
+        INPUT_PARAM_OPTION_QUOTE_PROC);
+    printf(
+        "\t%s, %s <filename>: generate an output file for a remote attestation "
+        "certificate, a report, or evidence.\n",
+        SHORT_INPUT_PARAM_OPTION_OUT_FILE,
+        INPUT_PARAM_OPTION_OUT_FILE);
+    printf(
+        "\t%s, %s <filename>: output a report in LEGACY_REPORT_REMOTE format "
+        "or evidence, and also its "
+        "endorsements binary.\n",
+        SHORT_INPUT_PARAM_OPTION_ENDORSEMENTS_FILENAME,
+        INPUT_PARAM_OPTION_ENDORSEMENTS_FILENAME);
+    printf(
+        "\t%s, %s: verify the generated remote attestation certificate, "
+        "report, "
+        "or evidence.\n",
+        SHORT_INPUT_PARAM_OPTION_VERIFY,
+        INPUT_PARAM_OPTION_VERIFY);
+    printf(
+        "\t%s, %s <filename>: generate a log file (default: %s).\n",
+        SHORT_INPUT_PARAM_OPTION_LOG_FILE,
+        INPUT_PARAM_OPTION_LOG_FILE,
+        DEFAULT_LOG_FILE);
+    printf("\t%s: enable verbose output.\n", INPUT_PARAM_OPTION_VERBOSE);
+    printf("Examples:\n");
+    printf("\t1. Show the verification results of evidence in SGX_ECDSA "
+           "format:\n");
+    printf("\t   oeutil generate-evidence -f sgx_ecdsa -v\n");
+    printf("\t2. Generate and output a certificate:\n");
+    printf("\t   oeutil generate -f cert private.pem public.pem -o cert.der\n");
+    printf("\t3. Generate and output a report:\n");
+    printf("\t   oeutil gen --format legacy_report_remote --out report.bin\n");
 }
 
 void dump_certificate(const uint8_t* data, size_t data_len)
@@ -87,16 +338,6 @@ void decode_certificate_pem(FILE* file, const uint8_t* data, size_t data_len)
             x509,
             XN_FLAG_COMPAT,
             XN_FLAG_SEP_CPLUS_SPC | XN_FLAG_DUMP_UNKNOWN_FIELDS);
-    BIO_free_all(input);
-}
-
-void decode_crl_pem(const uint8_t* data, size_t data_len)
-{
-    X509_CRL* x509;
-    BIO* input = BIO_new_mem_buf(data, (int)data_len);
-    x509 = PEM_read_bio_X509_CRL(input, NULL, NULL, NULL);
-    if (x509)
-        X509_CRL_print_fp(log_file, x509);
     BIO_free_all(input);
 }
 
@@ -264,62 +505,6 @@ void dump_claims(const oe_claim_t* claims, size_t claims_length)
     {
         printf("claims[%zu]: %s\n%s\n\n", i, claims[i].name, claims[i].value);
     }
-}
-
-// DCAP client (libdcap_quoteprov) log callback to this function.
-void oegenerate_quote_provider_log(
-    sgx_ql_log_level_t level,
-    const char* message)
-{
-    const char* level_string = level == 0 ? "ERROR" : "INFO";
-
-    log("[%s]: %s\n", level_string, message);
-}
-
-// Set DCAP client (libdcap_quoteprov) log callback
-void set_log_callback()
-{
-#if defined(__linux__)
-    extern oe_sgx_quote_provider_t provider;
-
-    // Initialize quote provider and set log callback
-    oe_initialize_quote_provider();
-
-    sgx_ql_set_logging_function_t set_log_fcn =
-        (sgx_ql_set_logging_function_t)dlsym(
-            provider.handle, "sgx_ql_set_logging_function");
-    if (set_log_fcn != nullptr)
-    {
-        set_log_fcn(oegenerate_quote_provider_log);
-    }
-#endif
-}
-
-OE_INLINE uint16_t read_uint16(const uint8_t* p)
-{
-    return (uint16_t)(p[0] | (p[1] << 8));
-}
-
-OE_INLINE uint32_t read_uint32(const uint8_t* p)
-{
-    return (uint32_t)(p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24));
-}
-
-oe_result_t output_file(
-    const char* file_name,
-    const uint8_t* data,
-    size_t data_size)
-{
-    FILE* output = nullptr;
-    fopen_s(&output, file_name, "wb");
-    if (!output)
-    {
-        log("Failed to open output file %s\n", file_name);
-        return OE_FAILURE;
-    }
-    fwrite(data, data_size, 1, output);
-    fclose(output);
-    return OE_OK;
 }
 
 oe_result_t dump_sgx_quote(
@@ -536,13 +721,27 @@ oe_result_t dump_oe_endorsements(
         oe_sgx_endorsement_item crl_pck_cert =
             endorsements.items[OE_SGX_ENDORSEMENT_FIELD_CRL_PCK_CERT];
         log("Endorsement: CRL PCK Certificate:\n");
-        decode_crl_pem(crl_pck_cert.data, crl_pck_cert.size);
+        if (crl_pck_cert.size >= OE_PEM_BEGIN_CRL_LEN &&
+            memcmp(
+                (const char*)crl_pck_cert.data,
+                OE_PEM_BEGIN_CRL,
+                OE_PEM_BEGIN_CRL_LEN) == 0)
+            decode_crl_pem(crl_pck_cert.data, crl_pck_cert.size);
+        else
+            decode_crl_der(crl_pck_cert.data, crl_pck_cert.size);
         log("\n");
 
         oe_sgx_endorsement_item crl_pck_proc_ca =
             endorsements.items[OE_SGX_ENDORSEMENT_FIELD_CRL_PCK_PROC_CA];
         log("Endorsement: CRL PCK Proc CA:\n");
-        decode_crl_pem(crl_pck_proc_ca.data, crl_pck_proc_ca.size);
+        if (crl_pck_proc_ca.size >= OE_PEM_BEGIN_CRL_LEN &&
+            memcmp(
+                (const char*)crl_pck_proc_ca.data,
+                OE_PEM_BEGIN_CRL,
+                OE_PEM_BEGIN_CRL_LEN) == 0)
+            decode_crl_pem(crl_pck_proc_ca.data, crl_pck_proc_ca.size);
+        else
+            decode_crl_der(crl_pck_proc_ca.data, crl_pck_proc_ca.size);
         log("\n");
 
         oe_sgx_endorsement_item crl_issuer_chain =
@@ -606,6 +805,56 @@ done:
     return result;
 }
 
+// This is the identity validation callback. A TLS connecting party (client or
+// server) can verify the passed in "identity" information to decide whether to
+// accept a connection request.
+oe_result_t enclave_identity_verifier(oe_identity_t* identity, void* arg)
+{
+    oe_result_t result = OE_VERIFY_FAILED;
+
+    (void)arg;
+    printf("enclave_identity_verifier is called with parsed report:\n");
+
+    // Check the enclave's security version
+    printf("identity.security_version = %d\n", identity->security_version);
+    if (identity->security_version < 1)
+    {
+        printf(
+            "identity.security_version check failed (%d)\n",
+            identity->security_version);
+        goto done;
+    }
+
+    // Dump an enclave's unique ID, signer ID and Product ID. They are
+    // MRENCLAVE, MRSIGNER and ISVPRODID for SGX enclaves.  In a real scenario,
+    // custom id checking should be done here
+    printf("identity->unique_id :\n");
+    for (int i = 0; i < OE_UNIQUE_ID_SIZE; i++)
+        printf("0x%0x ", (uint8_t)identity->unique_id[i]);
+
+    printf("\nidentity->signer_id :\n");
+    for (int i = 0; i < OE_SIGNER_ID_SIZE; i++)
+        printf("0x%0x ", (uint8_t)identity->signer_id[i]);
+
+    // verify signer id
+    OE_CHECK_MSG(
+        verify_signer_id(
+            (char*)OEUTIL_ENC_PUBLIC_KEY,
+            sizeof(OEUTIL_ENC_PUBLIC_KEY),
+            identity->signer_id,
+            sizeof(identity->signer_id)),
+        "Failed to verify signer id. Error: (%s)\n",
+        oe_result_str(result));
+
+    printf("\nidentity->product_id :\n");
+    for (int i = 0; i < OE_PRODUCT_ID_SIZE; i++)
+        printf("0x%0x ", (uint8_t)identity->product_id[i]);
+
+    result = OE_OK;
+done:
+    return result;
+}
+
 oe_result_t get_oe_report_from_certificate(
     const uint8_t* certificate_in_der,
     size_t certificate_in_der_length,
@@ -640,10 +889,96 @@ oe_result_t get_oe_report_from_certificate(
     }
     else
     {
-        if (!report_buffer)
-            free(report_buffer);
+        free(report_buffer);
     }
 
+    return result;
+}
+
+oe_result_t generate_certificate(
+    oe_enclave_t* enclave,
+    uint8_t* private_key,
+    size_t private_key_size,
+    uint8_t* public_key,
+    size_t public_key_size,
+    const char* certificate_filename,
+    bool verify,
+    bool verbose)
+{
+    oe_result_t result = OE_UNEXPECTED;
+    oe_result_t ecall_result;
+    unsigned char* certificate = nullptr;
+    size_t certificate_size = 0;
+    uint8_t* report = nullptr;
+    size_t report_size = 0;
+
+    log("========== Creating certificate with given private/public keys.\n");
+    result = get_tls_cert_signed_with_key(
+        enclave,
+        &ecall_result,
+        private_key,
+        private_key_size,
+        public_key,
+        public_key_size,
+        &certificate,
+        &certificate_size);
+
+    if ((result != OE_OK) || (ecall_result != OE_OK))
+    {
+        printf(
+            "Failed to create certificate. Enclave: %s, Host: %s\n",
+            oe_result_str(ecall_result),
+            oe_result_str(result));
+        goto done;
+    }
+    if (certificate_filename)
+    {
+        result =
+            output_file(certificate_filename, certificate, certificate_size);
+        if (result != OE_OK)
+        {
+            printf(
+                "Failed to open certificate file %s\n", certificate_filename);
+            goto done;
+        }
+    }
+
+    if (verbose)
+    {
+        printf("\n");
+        dump_certificate(certificate, certificate_size);
+
+        if (get_oe_report_from_certificate(
+                certificate, certificate_size, &report, &report_size) == OE_OK)
+        {
+            dump_oe_report(report, report_size);
+        }
+    }
+
+    log("========== Got certificate = %p certificate_size = %zu\n",
+        certificate,
+        certificate_size);
+
+    if (verify) // validate certificate
+    {
+        OE_CHECK_MSG(
+            oe_verify_attestation_certificate(
+                certificate,
+                certificate_size,
+                enclave_identity_verifier,
+                nullptr),
+            "Failed to verify certificate. result=%u (%s)\n",
+            result,
+            oe_result_str(result));
+
+        log("========== Certificate verified\n\n");
+    }
+
+done:
+    // deallcate resources
+    free(certificate);
+
+    fflush(stdout);
     return result;
 }
 
@@ -754,8 +1089,8 @@ oe_result_t generate_oe_report(
         // verify signer id
         OE_CHECK_MSG(
             verify_signer_id(
-                (char*)OEGENERATE_ENC_PUBLIC_KEY,
-                sizeof(OEGENERATE_ENC_PUBLIC_KEY),
+                (char*)OEUTIL_ENC_PUBLIC_KEY,
+                sizeof(OEUTIL_ENC_PUBLIC_KEY),
                 parsed_report.identity.signer_id,
                 sizeof(parsed_report.identity.signer_id)),
             "Failed to verify signer id. Error: (%s)\n",
@@ -768,8 +1103,7 @@ oe_result_t generate_oe_report(
     result = OE_OK;
 
 done:
-    if (remote_report)
-        oe_free_report(remote_report);
+    oe_free_report(remote_report);
 
     return result;
 }
@@ -926,8 +1260,8 @@ oe_result_t generate_oe_evidence(
         }
         OE_CHECK_MSG(
             verify_signer_id(
-                (char*)OEGENERATE_ENC_PUBLIC_KEY,
-                sizeof(OEGENERATE_ENC_PUBLIC_KEY),
+                (char*)OEUTIL_ENC_PUBLIC_KEY,
+                sizeof(OEUTIL_ENC_PUBLIC_KEY),
                 claim->value,
                 claim->value_size),
             "Failed to verify signer id. Error: (%s)\n",
@@ -951,4 +1285,480 @@ oe_result_t generate_oe_evidence(
 
 done:
     return result;
+}
+
+// Get full path of oeutil running executable, then get enclave filename by:
+// In linux, replace "<path>/oeutil" with "<path>/oeutil_enc.signed".
+// In windows, replace "<path>/oeutil.exe" with "<path>/oeutil_enc.signed".
+static char* _get_enclave_filename()
+{
+    char* enclave_filename = nullptr;
+    char path[OE_PATH_MAX];
+    size_t path_size = 0;
+    size_t enclave_filename_size = 0;
+
+#if defined(__linux__)
+    path_size += (size_t)readlink("/proc/self/exe", path, OE_PATH_MAX);
+#elif defined(_WIN32)
+    path_size += (size_t)GetModuleFileName(nullptr, path, OE_PATH_MAX);
+    path_size -= strlen(".exe");
+#endif
+
+    if (path_size < 0 || path_size >= OE_PATH_MAX)
+    {
+        printf("Failed to read enclave full path.\n");
+        goto done;
+    }
+    path[path_size] = '\0';
+    enclave_filename_size = path_size + sizeof(ENCLAVE_FILENAME_SUFFIX);
+    enclave_filename = (char*)malloc(enclave_filename_size);
+
+    if (enclave_filename == nullptr)
+    {
+        printf("Failed to malloc enclave_filename.\n");
+        goto done;
+    }
+    snprintf(
+        enclave_filename,
+        enclave_filename_size,
+        "%s%s",
+        path,
+        ENCLAVE_FILENAME_SUFFIX);
+
+    // Verify enclave file is valid
+    FILE* fp;
+    fopen_s(&fp, enclave_filename, "rb");
+    if (!fp)
+    {
+        printf("Enclave file does not exist: %s.\n", enclave_filename);
+        goto done;
+    }
+    else
+        fclose(fp);
+
+done:
+    return enclave_filename;
+}
+
+int _parse_args(int argc, const char* argv[])
+{
+    // clear parameters memory
+    memset(&_parameters, 0, sizeof(_parameters));
+
+    // save
+    _parameters.generate_certificate = false;
+    _parameters.generate_legacy_report_remote = false;
+    _parameters.generate_sgx_ecdsa = false;
+    _parameters.generate_sgx_epid_linkable = false;
+    _parameters.generate_sgx_epid_unlinkable = false;
+    _parameters.out_filename = nullptr;
+    _parameters.endorsements_filename = nullptr;
+    _parameters.log_filename = DEFAULT_LOG_FILE;
+    _parameters.quote_proc = "";
+    _parameters.verify = false;
+    _parameters.verbose = false;
+
+    int i = 2; // current index
+
+    if (argc == 3 && (strcasecmp(INPUT_PARAM_OPTION_HELP, argv[i]) == 0 ||
+                      strcasecmp(SHORT_INPUT_PARAM_OPTION_HELP, argv[i]) == 0))
+    {
+        _display_help(argv[0]);
+        return 0;
+    }
+
+    if (argc < 4)
+    {
+        _display_help(argv[0]);
+        return 1;
+    }
+
+    while (i < argc)
+    {
+        if (strcasecmp(INPUT_PARAM_OPTION_FORMAT, argv[i]) == 0 ||
+            strcasecmp(SHORT_INPUT_PARAM_OPTION_FORMAT, argv[i]) == 0)
+        {
+            if (argc < i + 2)
+                break;
+
+            if (strcasecmp(INPUT_PARAM_OPTION_CERT, argv[i + 1]) == 0)
+            {
+                if (argc < i + 4)
+                    break;
+
+                _parameters.generate_certificate = true;
+                _parameters.private_key_filename = argv[i + 2];
+                _parameters.public_key_filename = argv[i + 3];
+                i += 4;
+            }
+            else if (
+                strcasecmp(
+                    INPUT_PARAM_OPTION_LEGACY_REPORT_REMOTE, argv[i + 1]) == 0)
+            {
+                _parameters.generate_legacy_report_remote = true;
+                i += 2;
+            }
+            else if (strcasecmp(INPUT_PARAM_OPTION_SGX_ECDSA, argv[i + 1]) == 0)
+            {
+                _parameters.generate_sgx_ecdsa = true;
+                i += 2;
+            }
+            else if (
+                strcasecmp(INPUT_PARAM_OPTION_SGX_EPID_LINKABLE, argv[i + 1]) ==
+                0)
+            {
+                _parameters.generate_sgx_epid_linkable = true;
+                i += 2;
+            }
+            else if (
+                strcasecmp(
+                    INPUT_PARAM_OPTION_SGX_EPID_UNLINKABLE, argv[i + 1]) == 0)
+            {
+                _parameters.generate_sgx_epid_unlinkable = true;
+                i += 2;
+            }
+            else
+            {
+                printf("Invalid format: %s\n\n", argv[i + 1]);
+                break;
+            }
+        }
+        else if (
+            strcasecmp(INPUT_PARAM_OPTION_QUOTE_PROC, argv[i]) == 0 ||
+            strcasecmp(SHORT_INPUT_PARAM_OPTION_QUOTE_PROC, argv[i]) == 0)
+        {
+            if (argc < i + 2)
+                break;
+
+            _parameters.quote_proc = argv[i + 1];
+            if (strcasecmp(INPUT_PARAM_QUOTE_IN_PROC, _parameters.quote_proc) !=
+                    0 &&
+                strcasecmp(
+                    INPUT_PARAM_QUOTE_OUT_OF_PROC, _parameters.quote_proc) != 0)
+            {
+                printf(
+                    "Please use 'in' or 'out' with %s.\n",
+                    INPUT_PARAM_OPTION_QUOTE_PROC);
+                return 1;
+            }
+            i += 2;
+        }
+        else if (
+            strcasecmp(INPUT_PARAM_OPTION_ENDORSEMENTS_FILENAME, argv[i]) ==
+                0 ||
+            strcasecmp(
+                SHORT_INPUT_PARAM_OPTION_ENDORSEMENTS_FILENAME, argv[i]) == 0)
+        {
+            if (argc < i + 2)
+                break;
+
+            _parameters.endorsements_filename = argv[i + 1];
+            i += 2;
+        }
+        else if (
+            strcasecmp(INPUT_PARAM_OPTION_OUT_FILE, argv[i]) == 0 ||
+            strcasecmp(SHORT_INPUT_PARAM_OPTION_OUT_FILE, argv[i]) == 0)
+        {
+            if (argc < i + 2)
+                break;
+
+            _parameters.out_filename = argv[i + 1];
+            i += 2;
+        }
+        else if (
+            strcasecmp(INPUT_PARAM_OPTION_VERIFY, argv[i]) == 0 ||
+            strcasecmp(SHORT_INPUT_PARAM_OPTION_VERIFY, argv[i]) == 0)
+        {
+            _parameters.verify = true;
+            i++;
+        }
+        else if (
+            strcasecmp(INPUT_PARAM_OPTION_LOG_FILE, argv[i]) == 0 ||
+            strcasecmp(SHORT_INPUT_PARAM_OPTION_LOG_FILE, argv[i]) == 0)
+        {
+            if (argc < i + 2)
+                break;
+
+            _parameters.log_filename = argv[i + 1];
+            i += 2;
+        }
+        else if (strcasecmp(INPUT_PARAM_OPTION_VERBOSE, argv[i]) == 0)
+        {
+            _parameters.verbose = true;
+            i++;
+        }
+        else
+        {
+            printf("Invalid option: %s\n\n", argv[i]);
+            _display_help(argv[0]);
+            return 1;
+        }
+    }
+
+    if (i < argc)
+    {
+        printf("%s has invalid number of parameters.\n\n", argv[i]);
+        _display_help(argv[0]);
+        return 1;
+    }
+
+    if (_parameters.generate_certificate +
+            _parameters.generate_legacy_report_remote +
+            _parameters.generate_sgx_ecdsa +
+            _parameters.generate_sgx_epid_linkable +
+            _parameters.generate_sgx_epid_unlinkable !=
+        1)
+    {
+        printf("Please specify to generate a certificate, a report, or "
+               "evidence in SGX_ECDSA, SGX_EPID_LINKABLE or "
+               "SGX_EPID_UNLINKABLE format.\n");
+        return 1;
+    }
+
+    return 0;
+}
+
+oe_result_t _process_parameters(oe_enclave_t* enclave)
+{
+    oe_result_t result = OE_FAILURE;
+
+#if defined(__linux__)
+    char* sgx_aesm_env = getenv(SGX_AESM_ADDR);
+
+    // For Linux, if "SGX_AESM_ADDR" not set and out-of-proc is required, set
+    // "SGX_AESM_ADDR" to "1" and unset after process finishes
+    if (sgx_aesm_env == nullptr)
+    {
+        if (strcasecmp(INPUT_PARAM_QUOTE_OUT_OF_PROC, _parameters.quote_proc) ==
+                0 &&
+            setenv(SGX_AESM_ADDR, "1", 1) != 0)
+        {
+            printf("Failed to set environment variable 'SGX_AESM_ADDR'\n");
+            goto done;
+        }
+    }
+    // if "SGX_AESM_ADDR" is set and in-proc is required, unset it during the
+    // process and reset it to its original value after process finishes
+    else if (
+        strcasecmp(INPUT_PARAM_QUOTE_IN_PROC, _parameters.quote_proc) == 0 &&
+        unsetenv(SGX_AESM_ADDR) != 0)
+    {
+        printf("Failed to unset environment variable 'SGX_AESM_ADDR'\n");
+        goto done;
+    }
+#elif defined(_WIN32)
+    char sgx_aesm_env[SGX_AESM_ADDR_MAXSIZE];
+    int env_size = GetEnvironmentVariableA(
+        SGX_AESM_ADDR, sgx_aesm_env, SGX_AESM_ADDR_MAXSIZE);
+
+    if ((env_size == 0 && GetLastError() != ERROR_ENVVAR_NOT_FOUND) ||
+        env_size >= SGX_AESM_ADDR_MAXSIZE)
+    {
+        printf("Failed to read environment variable 'SGX_AESM_ADDR'\n");
+        goto done;
+    }
+
+    // For Windows, out-of-proc is not tested as extra dependencies required
+    if (strcasecmp(INPUT_PARAM_QUOTE_OUT_OF_PROC, _parameters.quote_proc) == 0)
+    {
+        printf("In-proc quoting is by default on Windows. Please use in-proc "
+               "quoting\n");
+        goto done;
+    }
+    // if "SGX_AESM_ADDR" is set and in-proc is required, unset it during the
+    // process and reset it to its original value after process finishes
+    else if (
+        strcasecmp(INPUT_PARAM_QUOTE_IN_PROC, _parameters.quote_proc) == 0 &&
+        env_size != 0 && SetEnvironmentVariableA(SGX_AESM_ADDR, nullptr) == 0)
+    {
+        printf("Failed to unset environment variable 'SGX_AESM_ADDR'\n");
+        goto done;
+    }
+#endif
+
+    if (_parameters.generate_certificate)
+    {
+        size_t private_key_size;
+        uint8_t* private_key;
+        size_t public_key_size;
+        uint8_t* public_key;
+
+        // read private key (pem format)
+        OE_CHECK(_read_key(
+            _parameters.private_key_filename, &private_key, &private_key_size));
+        OE_CHECK(_read_key(
+            _parameters.public_key_filename, &public_key, &public_key_size));
+        OE_CHECK(generate_certificate(
+            enclave,
+            private_key,
+            private_key_size,
+            public_key,
+            public_key_size,
+            _parameters.out_filename,
+            _parameters.verify,
+            _parameters.verbose));
+    }
+    else if (_parameters.generate_legacy_report_remote)
+    {
+        OE_CHECK(generate_oe_report(
+            enclave,
+            _parameters.out_filename,
+            _parameters.endorsements_filename,
+            _parameters.verify,
+            _parameters.verbose));
+    }
+    else if (_parameters.generate_sgx_ecdsa)
+    {
+        OE_CHECK(generate_oe_evidence(
+            enclave,
+            _sgx_ecdsa_uuid,
+            _parameters.out_filename,
+            _parameters.endorsements_filename,
+            _parameters.verify,
+            _parameters.verbose));
+    }
+    else if (_parameters.generate_sgx_epid_linkable)
+    {
+        OE_CHECK(generate_oe_evidence(
+            enclave,
+            _sgx_epid_linkable_uuid,
+            _parameters.out_filename,
+            _parameters.endorsements_filename,
+            _parameters.verify,
+            _parameters.verbose));
+    }
+    else if (_parameters.generate_sgx_epid_unlinkable)
+    {
+        OE_CHECK(generate_oe_evidence(
+            enclave,
+            _sgx_epid_unlinkable_uuid,
+            _parameters.out_filename,
+            _parameters.endorsements_filename,
+            _parameters.verify,
+            _parameters.verbose));
+    }
+
+    result = OE_OK;
+
+done:
+
+#if defined(__linux__)
+    if (sgx_aesm_env == nullptr)
+    {
+        if (strcasecmp(INPUT_PARAM_QUOTE_OUT_OF_PROC, _parameters.quote_proc) ==
+                0 &&
+            unsetenv(SGX_AESM_ADDR) != 0)
+        {
+            printf(
+                "Failed to unset environment variable 'SGX_AESM_ADDR', please "
+                "manually unset it\n");
+            result = OE_FAILURE;
+        }
+    }
+    else if (
+        strcasecmp(INPUT_PARAM_QUOTE_IN_PROC, _parameters.quote_proc) == 0 &&
+        setenv(SGX_AESM_ADDR, sgx_aesm_env, 1) != 0)
+    {
+        printf(
+            "Failed to reset environment variable 'SGX_AESM_ADDR', please "
+            "manually reset it as %s\n",
+            sgx_aesm_env);
+        result = OE_FAILURE;
+    }
+#elif defined(_WIN32)
+    if (env_size != 0 &&
+        strcasecmp(INPUT_PARAM_QUOTE_IN_PROC, _parameters.quote_proc) == 0)
+    {
+        if (SetEnvironmentVariableA(SGX_AESM_ADDR, sgx_aesm_env) == 0)
+        {
+            printf(
+                "Failed to reset environment variable 'SGX_AESM_ADDR', please "
+                "manually reset it as %s\n",
+                sgx_aesm_env);
+            result = OE_FAILURE;
+        }
+    }
+#endif
+    return result;
+}
+
+int oeutil_generate_evidence(int argc, const char* argv[])
+{
+    int ret = 0;
+    printf("NOTICE: oeutil generate-evidence is purely a debugging utility and "
+           "not suitable for production use.\n\n");
+    if (!oe_has_sgx_quote_provider())
+    {
+        fprintf(
+            stderr, "FAILURE: DCAP libraries must be present for this test.\n");
+        return 1;
+    }
+
+    oe_result_t result;
+    oe_enclave_t* enclave = nullptr;
+    char* enclave_filename = nullptr;
+
+    const uint32_t flags = oe_get_create_flags();
+    if ((flags & OE_ENCLAVE_FLAG_SIMULATE) != 0)
+    {
+        printf(
+            "oeutil generate-evidence is not supported in simulation mode.\n");
+        goto done;
+    }
+
+    enclave_filename = _get_enclave_filename();
+    if (enclave_filename == nullptr)
+    {
+        printf("Failed to get enclave filename.\n");
+        goto done;
+    }
+
+    ret = _parse_args(argc, argv);
+    if (ret != 0)
+        goto done;
+
+    if ((result = oe_create_oeutil_enclave(
+             enclave_filename,
+             OE_ENCLAVE_TYPE_AUTO,
+             OE_ENCLAVE_FLAG_DEBUG,
+             nullptr,
+             0,
+             &enclave)) != OE_OK)
+    {
+        printf(
+            "Failed to create enclave. result=%u (%s)\n",
+            result,
+            oe_result_str(result));
+        ret = 1;
+        goto done;
+    }
+
+    // Create log file
+#ifdef _WIN32
+    fopen_s(&log_file, _parameters.log_filename, "w");
+#else
+    log_file = fopen(_parameters.log_filename, "w");
+#endif
+    if (!log_file)
+    {
+        printf("Failed to open log file %s\n", _parameters.log_filename);
+        ret = 1;
+        goto done;
+    }
+
+    set_log_callback();
+
+    if ((result = _process_parameters(enclave)) != OE_OK)
+    {
+        printf(
+            "Failed to process parameters. Error: %s\n", oe_result_str(result));
+        ret = 1;
+        goto done;
+    }
+
+    result = oe_terminate_enclave(enclave);
+done:
+
+    free(enclave_filename);
+    return ret;
 }
