+++ tools/oeutil/host/parse_args_helper.h
@@ -1,0 +1,7 @@
+// Copyright (c) Open Enclave SDK contributors.
+// Licensed under the MIT License.
+
+#if defined(_WIN32)
+#define strcasecmp stricmp
+#define strncasecmp strnicmp
+#endif
